[긴토키] 신규 능력 구현 및 시스템 개선 (에미야 시로 고도화 포함)
- 기능 작업 -
기존 기능 수정
   ㄴ Ability.java - 소수점(double) 단위 정교한 쿨타임 설정 기능 구현
   ㄴ AbilityManager.java - 신규 능력 긴토키(Gintoki) 등록

- 능력 작업 -
신규 능력 구현
   ㄴ 긴토키 - 네오 암스트롱 사이클론 제트 암스트롱 포 패턴 구현 및 발사 로직 최적화
   ㄴ 긴토키 - 쿨타임 상향(1.5초), 바닐라 설치 차단, 자가 피해 적용 등 완성도 강화
기존 능력 수정
   ㄴ 에미야 시로 - 검 소환 시 파란색 파티클 효과 추가
   ㄴ 에미야 시로 - 발사된 검 10초 후 자동 제거(TTL) 로직 추가로 성능 최적화

+++ src/main/java/me/user/moc/ability/Ability.java
@@ -90,8 +90,8 @@ public abstract class Ability implements Listener {
     }
 
     // === [쿨타임 관련 헬퍼 메서드] ===
-    protected void setCooldown(Player p, long seconds) {
-        cooldowns.put(p.getUniqueId(), System.currentTimeMillis() + (seconds * 1000));
+    protected void setCooldown(Player p, double seconds) {
+        cooldowns.put(p.getUniqueId(), System.currentTimeMillis() + (long) (seconds * 1000));
 
         // [추가] 기존 알림 취소
         if (cooldownNotifyTasks.containsKey(p.getUniqueId())) {
@@ -108,7 +108,7 @@ public abstract class Ability implements Listener {
                 }
                 cooldownNotifyTasks.remove(p.getUniqueId());
             }
-        }.runTaskLater(plugin, seconds * 20L);
+        }.runTaskLater(plugin, (long) (seconds * 20));
 
         cooldownNotifyTasks.put(p.getUniqueId(), task);
     }
